+++ platforms/nrf5/libraries/memfault/platform_reference_impl/memfault_platform_core.c
@@ -6,8 +6,25 @@
 
 #include "app_util_platform.h"
 #include "memfault/core/compiler.h"
+#include "memfault/core/debug_log.h"
+#include "memfault/panics/assert.h"
+#include "memfault/panics/coredump.h"
+
+//! Sanity check that coredump storage is large enough to capture the regions we want to collect
+static void prv_assert_coredump_storage_appropriately_sized(void) {
+  sMfltCoredumpStorageInfo storage_info = { 0 };
+  memfault_platform_coredump_storage_get_info(&storage_info);
+  const size_t size_needed = memfault_coredump_storage_compute_size_required();
+  if (size_needed > storage_info.size) {
+    MEMFAULT_LOG_ERROR("Coredump storage too small. Got %d B, need %d B",
+                       storage_info.size, size_needed);
+  }
+  MEMFAULT_ASSERT(size_needed <= storage_info.size);
+}
 
 int memfault_platform_boot(void) {
+  MEMFAULT_LOG_INFO("Memfault demo app starting!");
+  prv_assert_coredump_storage_appropriately_sized();
   return 0;
 }